Image Customizer: Split up customizeutils.go.
Split up the functions in 'customizeutils.go` into separate files for
each feature. This should hopefully make the code a little easier to
navigate, particularly in regard to tests.

No code was changed in this refactor. Just movement between files.
