+++ toolkit/tools/pkg/imagecustomizerlib/customizefiles.go
@@ -1,0 +1,70 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+package imagecustomizerlib
+
+import (
+	"fmt"
+	"io/fs"
+
+	"github.com/microsoft/azurelinux/toolkit/tools/imagecustomizerapi"
+	"github.com/microsoft/azurelinux/toolkit/tools/internal/file"
+	"github.com/microsoft/azurelinux/toolkit/tools/internal/logger"
+	"github.com/microsoft/azurelinux/toolkit/tools/internal/safechroot"
+)
+
+const (
+	defaultFilePermissions = 0o755
+)
+
+func copyAdditionalFiles(baseConfigPath string, additionalFiles imagecustomizerapi.AdditionalFilesMap, imageChroot *safechroot.Chroot) error {
+	for sourceFile, fileConfigs := range additionalFiles {
+		absSourceFile := file.GetAbsPathWithBase(baseConfigPath, sourceFile)
+		for _, fileConfig := range fileConfigs {
+			logger.Log.Infof("Copying: %s", fileConfig.Path)
+
+			fileToCopy := safechroot.FileToCopy{
+				Src:         absSourceFile,
+				Dest:        fileConfig.Path,
+				Permissions: (*fs.FileMode)(fileConfig.Permissions),
+			}
+
+			err := imageChroot.AddFiles(fileToCopy)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
+func copyAdditionalDirs(baseConfigPath string, additionalDirs imagecustomizerapi.DirConfigList, imageChroot *safechroot.Chroot) error {
+	for _, dirConfigElement := range additionalDirs {
+		absSourceDir := file.GetAbsPathWithBase(baseConfigPath, dirConfigElement.SourcePath)
+		logger.Log.Infof("Copying %s to %s", absSourceDir, dirConfigElement.DestinationPath)
+
+		// Setting permissions values. They are set to a default value if they have not been specified.
+		newDirPermissionsValue := fs.FileMode(defaultFilePermissions)
+		if dirConfigElement.NewDirPermissions != nil {
+			newDirPermissionsValue = *(*fs.FileMode)(dirConfigElement.NewDirPermissions)
+		}
+		childFilePermissionsValue := fs.FileMode(defaultFilePermissions)
+		if dirConfigElement.ChildFilePermissions != nil {
+			childFilePermissionsValue = *(*fs.FileMode)(dirConfigElement.ChildFilePermissions)
+		}
+
+		dirToCopy := safechroot.DirToCopy{
+			Src:                  absSourceDir,
+			Dest:                 dirConfigElement.DestinationPath,
+			NewDirPermissions:    newDirPermissionsValue,
+			ChildFilePermissions: childFilePermissionsValue,
+			MergedDirPermissions: (*fs.FileMode)(dirConfigElement.MergedDirPermissions),
+		}
+		err := imageChroot.AddDirs(dirToCopy)
+		if err != nil {
+			return fmt.Errorf("failed to copy directory (%s) to (%s):\n%w", absSourceDir, dirConfigElement.DestinationPath, err)
+		}
+	}
+	return nil
+}
