+++ toolkit/tools/pkg/imagecustomizerlib/customizehostname.go
@@ -1,0 +1,29 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+package imagecustomizerlib
+
+import (
+	"fmt"
+	"path/filepath"
+
+	"github.com/microsoft/azurelinux/toolkit/tools/internal/file"
+	"github.com/microsoft/azurelinux/toolkit/tools/internal/logger"
+	"github.com/microsoft/azurelinux/toolkit/tools/internal/safechroot"
+)
+
+func UpdateHostname(hostname string, imageChroot safechroot.ChrootInterface) error {
+	if hostname == "" {
+		return nil
+	}
+
+	logger.Log.Infof("Setting hostname (%s)", hostname)
+
+	hostnameFilePath := filepath.Join(imageChroot.RootDir(), "etc/hostname")
+	err := file.Write(hostname, hostnameFilePath)
+	if err != nil {
+		return fmt.Errorf("failed to write hostname file: %w", err)
+	}
+
+	return nil
+}
